moved sample markets to different file
